scan, display transactions

+++ src/db/seed/accounts.ts
@@ -2,16 +2,21 @@ import { DB_NAME } from "@/db/db_name"
 import { accountTable, currencyTable } from "../schemas"
 import { openDatabaseSync } from "expo-sqlite"
 import { drizzle } from "drizzle-orm/expo-sqlite"
-import { createId } from "@paralleldrive/cuid2"
 
 export async function seedAccounts() {
   const sqliteDb = openDatabaseSync(DB_NAME)
   const db = drizzle(sqliteDb)
 
+  const existingAccounts = await db.select().from(accountTable).limit(1)
+  const existingCurrencies = await db.select().from(currencyTable).limit(1)
+  if (existingAccounts.length > 0 && existingCurrencies.length > 0) {
+    return false
+  }
+
   await db.delete(accountTable)
   await db.delete(currencyTable)
 
-  const currencyId = createId()
+  const currencyId = 1
 
   await db.insert(currencyTable).values([
     {
@@ -26,8 +31,10 @@ export async function seedAccounts() {
       name: "Default",
       balance: 0,
       color: "gray",
-      icon: "💰",
+      emoji: "💰",
       currencyId: currencyId,
     },
   ])
+
+  return true
 }
